Revert commit 246677c and 01d1cd1
We are making this change because this affects all computers that get
the compozed image. Disabling mouse and arrow keys should be disabled on
a per computer/user basis and is unfriendly to new users of vim. In
addition, it can make pairing with someone using vim more difficult as a
non-vim user, whereas if macvim continues to work as a normal editor, it
won't interrupt a pair.

- Feel free to override the default options by adding these lines to ~/.vimrc.local

```
noremap <Up> <NOP>
noremap <Down> <NOP>
noremap <Left> <NOP>
noremap <Right> <NOP>

set mouse=
```

Signed-off-by: Travis Yoder <trayoda@gmail.com>